+++ main.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/coaxial/tizinger/extractor"
+	"github.com/coaxial/tizinger/fip"
 )
 
 func main() {
-	var fipExtractor extractor.FipExtractor
+	var fipExtractor fip.Extractor
 	// TODO use two days  ago at midnight
 	list, _ := fipExtractor.Playlist(time.Now().Unix())
 	fmt.Printf("%#v", list)
